revert using shouldKeep in removeItem
- forgot shouldKeep attempts to place item in invo, where removeItem checks all storage locations for a spot
- add prettyPrint prototype
- create Polyfills.js to keep PrototypeOverrides cleaner
